Reduced test time
